+++ lib/src/ast.dart
@@ -12,6 +12,19 @@ class SourcePosition {
   int get length => end - start;
 }
 
+/// Provides an interface for generic query evaluation.
+abstract class QueryEvaluator<R> {
+  R evalText(TextQuery query);
+  R evalPhrase(PhraseQuery query);
+  R evalScope(ScopeQuery query);
+  R evalCompare(CompareQuery query);
+  R evalRange(RangeQuery query);
+  R evalNot(NotQuery query);
+  R evalGroup(GroupQuery query);
+  R evalAnd(AndQuery query);
+  R evalOr(OrQuery query);
+}
+
 /// Base interface for queries.
 abstract class Query {
   const Query({
@@ -35,6 +48,8 @@ abstract class Query {
   ///
   /// If the [Query] cannot be cast to [R] it will throw an exception.
   R cast<R extends Query>() => this as R;
+
+  R eval<R>(QueryEvaluator<R> evaluator);
 }
 
 /// Text query to match [text].
@@ -44,6 +59,9 @@ class TextQuery extends Query {
     required this.text,
     required super.position,
   });
+
+  @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalText(this);
 
   @override
   String toString({bool debug = false}) => _debug(debug, text);
@@ -59,12 +77,15 @@ class PhraseQuery extends TextQuery {
   });
 
   @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalPhrase(this);
+
+  @override
   String toString({bool debug = false}) =>
       '"${children.map((n) => n.toString(debug: debug)).join(' ')}"';
 }
 
 /// Scopes [child] [Query] to be applied only on the [field].
-@Deprecated('Use FieldScopeQuery instead.')
+@Deprecated('Use ScopeQuery instead.')
 class FieldScope extends Query {
   final TextQuery field;
   final Query child;
@@ -76,21 +97,29 @@ class FieldScope extends Query {
   });
 
   @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator
+      .evalScope(ScopeQuery(field: field, child: child, position: position));
+
+  @override
   String toString({bool debug = false}) =>
       '$field:${child.toString(debug: debug)}';
 }
 
 /// Scopes [child] [Query] to be applied only on the [field].
 // ignore: deprecated_member_use_from_same_package
-class FieldScopeQuery extends FieldScope {
-  const FieldScopeQuery({
+class ScopeQuery extends FieldScope {
+  const ScopeQuery({
     required super.field,
     required super.child,
     required super.position,
   });
+
+  @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalScope(this);
 }
 
 /// Describes a [field] [operator] [text] tripled (e.g. year < 2000).
+@Deprecated('Use CompareQuery instead.')
 class FieldCompareQuery extends Query {
   final TextQuery field;
   final TextQuery operator;
@@ -104,8 +133,26 @@ class FieldCompareQuery extends Query {
   });
 
   @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalCompare(CompareQuery(
+      field: field, operator: operator, text: text, position: position));
+
+  @override
   String toString({bool debug = false}) =>
       _debug(debug, '$field$operator$text');
+}
+
+/// Describes a [field] [operator] [text] tripled (e.g. year < 2000).
+// ignore: deprecated_member_use_from_same_package
+class CompareQuery extends FieldCompareQuery {
+  CompareQuery({
+    required super.field,
+    required super.operator,
+    required super.text,
+    required super.position,
+  });
+
+  @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalCompare(this);
 }
 
 /// Describes a range query between [start] and [end].
@@ -122,6 +169,9 @@ class RangeQuery extends Query {
     this.startInclusive = true,
     this.endInclusive = true,
   });
+
+  @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalRange(this);
 
   @override
   String toString({bool debug = false}) => _debug(
@@ -143,6 +193,9 @@ class NotQuery extends Query {
   });
 
   @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalNot(this);
+
+  @override
   String toString({bool debug = false}) => '-${child.toString(debug: debug)}';
 }
 
@@ -153,6 +206,9 @@ class GroupQuery extends Query {
     required this.child,
     required super.position,
   });
+
+  @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalGroup(this);
 
   @override
   String toString({bool debug = false}) => '(${child.toString(debug: debug)})';
@@ -168,6 +224,9 @@ class AndQuery extends Query {
   });
 
   @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalAnd(this);
+
+  @override
   String toString({bool debug = false}) =>
       '(${children.map((n) => n.toString(debug: debug)).join(' ')})';
 }
@@ -180,6 +239,9 @@ class OrQuery extends Query {
     required this.children,
     required super.position,
   });
+
+  @override
+  R eval<R>(QueryEvaluator<R> evaluator) => evaluator.evalOr(this);
 
   @override
   String toString({bool debug = false}) =>